+++ models/frame.py
@@ -12,7 +12,7 @@ class Frame(BaseModel, Base):
     """To create frame table"""
     __tablename__ = "frames"
     id = Column(Integer, primary_key=True)
-    stock_id = Column(Integer(6), ForeignKey("stocks.id"), nullable=False)
+    stock_id = Column(Integer, ForeignKey("stocks.id"), nullable=False)
     consultation_id = Column(Integer(6), ForeignKey("consultations.id"), nullable=False)
     name = Column(String(60), nullable=False)
     shape = Column(String(60), nullable=True)
